Extract logic for running an RPC server.
