+++ node/src/components/rpc_server/rpcs.rs
@@ -11,14 +11,20 @@ pub mod info;
 pub mod speculative_exec;
 pub mod state;
 
-use std::{str, sync::Arc};
+use std::{convert::Infallible, str, sync::Arc, time::Duration};
 
 use async_trait::async_trait;
+use http::header::ACCEPT_ENCODING;
+use hyper::server::{conn::AddrIncoming, Builder};
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
+use tokio::sync::oneshot;
+use tower::ServiceBuilder;
+use tracing::info;
+use warp::Filter;
 
-use casper_json_rpc::{Error, RequestHandlersBuilder, ReservedErrorCode};
+use casper_json_rpc::{Error, RequestHandlers, RequestHandlersBuilder, ReservedErrorCode};
 use casper_types::ProtocolVersion;
 
 use super::{ReactorEventT, RpcRequest};
@@ -225,6 +231,45 @@ pub(super) trait RpcWithOptionalParams {
         api_version: ProtocolVersion,
         params: Option<Self::OptionalRequestParams>,
     ) -> Result<Self::ResponseResult, Error>;
+}
+
+/// Start JSON RPC server in a background.
+pub(super) async fn run(
+    builder: Builder<AddrIncoming>,
+    handlers: RequestHandlers,
+    qps_limit: u64,
+    max_body_bytes: u32,
+    api_path: &'static str,
+    server_name: &'static str,
+) {
+    let make_svc = hyper::service::make_service_fn(move |_| {
+        let service_routes = casper_json_rpc::route(api_path, max_body_bytes, handlers.clone());
+
+        // Supports content negotiation for gzip responses. This is an interim fix until
+        // https://github.com/seanmonstar/warp/pull/513 moves forward.
+        let service_routes_gzip = warp::header::exact(ACCEPT_ENCODING.as_str(), "gzip")
+            .and(service_routes.clone())
+            .with(warp::compression::gzip());
+
+        let service = warp::service(service_routes_gzip.or(service_routes));
+        async move { Ok::<_, Infallible>(service.clone()) }
+    });
+
+    let make_svc = ServiceBuilder::new()
+        .rate_limit(qps_limit, Duration::from_secs(1))
+        .service(make_svc);
+
+    let server = builder.serve(make_svc);
+    info!(address = %server.local_addr(), "started {} server", server_name);
+
+    let (shutdown_sender, shutdown_receiver) = oneshot::channel::<()>();
+    let server_with_shutdown = server.with_graceful_shutdown(async {
+        shutdown_receiver.await.ok();
+    });
+
+    let _ = tokio::spawn(server_with_shutdown).await;
+    let _ = shutdown_sender.send(());
+    info!("{} server shut down", server_name);
 }
 
 #[cfg(test)]
